Outsourced ParsonsPuzzle submitSolution function
